Support for Write service: now supporting more data types; Some refactoring of API - added constrictors to simplify creation of structure instances for typical use cases; Added struct EncodedExtensionObject to encapsulate generic custom extension objects whose internal structure is opaque for the SDK (passed to the user application in encoded way).

+++ include/opcua/DataBuffer.h
@@ -27,7 +27,7 @@ namespace OWA {
         return buffer.data();
       }
       uint32_t size() {
-        return buffer.size();
+        return (uint32_t) buffer.size();
       }
       void sgetn(uint8_t* m, uint32_t count) {
 				if ((position + count) > buffer.size())
